List packages in CRAN incoming queue.
Source: R/cran_queue.R

Check where your package stands in the CRAN incoming queue.

cran_incoming(pkg = NULL, folders = c("archive", "inspect", "pending",
  "pretest", "publish", "recheck", "waiting"))

Arguments

pkg
Optionally provide a vector of package name to limit the results to these packages.

folders
Which folders of the CRAN FTP do you want to inspect? Default all the non-human folders.

Value

A tibble with the following columns:

the name of the package: package

the version of the package: version

the name of the folder where the package was found: cran_folder

Details

When submitting a package to CRAN, it undergoes a series of checks before it is published and publicly available. cran_incoming() allows you to check the packages that are currently in the queue, and the folder where they are located. This information could help you track your package submission. Only the following folders are considered: archive, inspect, noemail, pending, pretest, publish, recheck, waiting. The folders named after the initials of the CRAN volunteers are not inspected.

Disclaimer

The information provided here is only to give you an indication of where your package stands in the submission process. It can be useful to confirm that your package has been correctly uploaded to CRAN. Please consult the CRAN RepositoryPolicy if you have any questions.

Examples

# NOT RUN {
  ## all the packages in the CRAN incoming queue
  cran_incoming()
  ## if the package `foo` is in the queue, it will appear below
  cran_incoming(pkg = "foo")
# }