CRAN check results at the terminal

foghorn brings the results of the CRAN checks into your R terminal. This package is intended for developers that have published packages on CRAN. The results of the checks can be summarized as a tibble (see check_cran_results), or as “pretty” terminal output (see summary_cran_results). The package also provides functions to view the logs of the CRAN check results (see show_cran_results), or to visit the web pages hosted by CRAN (see visit_cran_check).