Show the CRAN check results for a package

Given a package name published on CRAN, print the outcomes of checks that return notes, warnings or errors, and optionally print the messages.

show_cran_results(pkg, show_log = TRUE, src = c("website", "crandb"), ...)

Arguments

pkg
name of the package on CRAN

show_log
Should the messages of the “Check Details” be printed? (logical)

src
if "website" the data is scrapped from the CRAN website, if "crandb" the data is downloaded from a RDS file hosted on the CRAN servers (which is used to generate the information found on the CRAN website).

...
additional arguments to control where the data from the check results are coming from and how they are downloaded from the CRAN servers (see Details section).

Value

NULL, used for its side effect of printing the CRAN messages

Details

Where does the data come from?

The data comes from the CRAN servers. They generate RDS files that contains information regarding the results of the checks for all the packages, and all the flavors. This data is then used to generate the web pages. foghorn provides access to either of these data sources. If you choose src = "website" the data is scrapped from the CRAN website. If you only need to check a few packages, this is a good option. If you choose src = "crandb" the RDS files (about 20Mb) are downloaded first from the CRAN servers.

When choosing src = "crandb" you can also specify the following options:

dest a folder where to store the RDS files (tempdir() by default).

protocol either https (default) or http.

overwrite when FALSE (default), if the file exists in dest then it will not be downloaded again. when TRUE the file gets downloaded everytime it's needed.

... additional arguments that will be passed to GET function to control how the file will be downloaded.

See also

Note that the tools package contains unexported functions that can be used to extract summary information from the check results. Specifically tools:::sumarize_CRAN_check_status is similar to show_cran_results.