Visit the CRAN check results page

Visit the page in your web browser for a given package or a maintainer's email address

Usage

visit_cran_check(pkg = NULL, email = NULL)

Arguments

pkg

name of the package to check the results for

email

email address of the package maintainer

Value

The URL from the CRAN check results page invisibly