cran_details {foghorn}	R Documentation

Get details about the CRAN check results for packages

Description

Given the names of packages published on CRAN, return the output of checks that return notes, warnings or errors.

Usage


cran_details(pkg, src = c("website", "crandb"), ...)

## S3 method for class 'cran_details'
summary(object, show_log = TRUE, print_ok = TRUE, ...)

summary_cran_details(
  pkg,
  src = c("website", "crandb"),
  show_log = TRUE,
  print_ok = TRUE,
  ...
)

Arguments

pkg
character vector of the names for the packages on CRAN

src
if "website" the data is scrapped from the CRAN website, if "crandb" the data is downloaded from a RDS file hosted on the CRAN servers (which is used to generate the information found on the CRAN website).

...
additional arguments to control where the data from the check results are coming from and how they are downloaded from the CRAN servers (see Details section).

object
an object created by cran_details

show_log
Should the messages of the “Check Details” be printed? (logical)

print_ok
if TRUE the summary method will print a "all clear" message for package(s) that have an OK status for all CRAN checks.

Details

Where does the data come from?

The data comes from the CRAN servers. They generate RDS files that contains information regarding the results of the checks for all the packages, and all the flavors. This data is then used to generate the web pages.

foghorn provides access to either of these data sources. If you choose src = "website" the data is scrapped from the CRAN website. If you only need to check a few packages, this is a good option. If you choose src = "crandb" the RDS files (about 20Mb) are downloaded first from the CRAN servers.

When choosing src = "crandb" you can also specify the following options:

dest a folder where to store the RDS files (tempdir() by default).

protocol either https (default) or http.

overwrite when FALSE (default), if the file exists in dest then it will not be downloaded again. When TRUE the file gets downloaded every time it's needed.

... additional arguments that will be passed to GET function to control how the file will be downloaded.

Value

a tibble listing the names of the packages that have non- OK check results, the nature of the result (WARN, ERROR, FAIL, NOTE, or other issues).

See Also

Note that the tools package contains unexported functions that can be used to extract summary information from the check results. Specifically tools:::sumarize_CRAN_check_status is similar to show_cran_results.

[Package foghorn version 1.2.3 Index]