visit_cran_check {foghorn}	R Documentation

Visit the CRAN check results page

Description

Visit the page in your web browser for a given package or a maintainer's email address

Usage


visit_cran_check(pkg = NULL, email = NULL)

Arguments

pkg
name of the package to check the results for

email
email address of the package maintainer

Value

The URL from the CRAN check results page invisibly

[Package foghorn version 1.2.3 Index]